+++ test/fixtures/sort_constructors_first/mixed_members.input.dart
@@ -1,0 +1,11 @@
+class Timer {
+  void start() {}
+
+  int count = 0;
+
+  void stop() {}
+
+  Timer();
+
+  void reset() => count = 0;
+}