+++ test/fixtures/sort_constructors_first/multiple_classes.input.dart
@@ -1,0 +1,11 @@
+class A {
+  void run() {}
+
+  A();
+}
+
+class B {
+  int value = 0;
+
+  B(this.value);
+}