+++ test/fixtures/sort_constructors_first/multiple_constructors_scattered.input.dart
@@ -1,0 +1,7 @@
+class Color {
+  void describe() => print('rgb');
+
+  Color(int r, int g, int b);
+
+  Color.black() : this(0, 0, 0);
+}